pyproject.toml
[project]
name = "immersion_whisper"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "faster-whisper>=1.1.1",
    "ffmpeg-python>=0.2.0",
    "fr-dep-news-trf",
    "genanki>=0.13.1",
    "pandas>=2.3.1",
    "peewee>=3.18.2",
    "pydantic>=2.11.7",
    "pydub>=0.25.1",
    "pysrt>=1.1.2",
    "pyyaml>=6.0.2",
    "spacy>=3.8.7",
]

[project.scripts]
imw = "immersion_whisper.main:main"

[tool.uv.sources]
fr-dep-news-trf = { url = "https://github.com/explosion/spacy-models/releases/download/fr_dep_news_trf-3.8.0/fr_dep_news_trf-3.8.0-py3-none-any.whl" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
